Split sessions by day, and into topical subsessions

+++ aclpub2/generate.py
@@ -13,12 +13,13 @@ PARENT_DIR = Path(__file__).parent
 def generate(*, path: str, proceedings: bool, handbook: bool, overwrite: bool):
     root = Path(path)
     build_dir = Path("build")
-    shutil.rmtree(build_dir, ignore_errors=True)
     build_dir.mkdir(exist_ok=True)
 
     # Throw if the build directory isn't empty, and the user did not specify an overwrite.
     if len([build_dir.iterdir()]) > 0 and not overwrite:
-        raise Exception(f"Build directory {build_dir} is not empty, and the overwrite flag is false.")
+        raise Exception(
+            f"Build directory {build_dir} is not empty, and the overwrite flag is false."
+        )
 
     # Load and preprocess the .yml configuration.
     (
@@ -33,7 +34,7 @@ def generate(*, path: str, proceedings: bool, handbook: bool, overwrite: bool):
     ) = load_configs(root)
     id_to_paper, alphabetized_author_index = process_papers(papers, root, proceedings)
     if proceedings:
-        sessions_by_date = process_program_by_page(program)
+        sessions_by_date = process_program_proceedings(program)
         template = load_template("proceedings")
         rendered_template = template.render(
             root=str(root),
@@ -56,7 +57,7 @@ def generate(*, path: str, proceedings: bool, handbook: bool, overwrite: bool):
 
     if handbook:
         template = load_template("handbook")
-        program = process_program(program)
+        program = process_program_handbook(program)
         rendered_template = template.render(
             root=str(root),
             conference=conference,
@@ -69,12 +70,13 @@ def generate(*, path: str, proceedings: bool, handbook: bool, overwrite: bool):
             papers=papers,
             id_to_paper=id_to_paper,
             program=program,
-            build_dir=str(build_dir)
+            build_dir=str(build_dir),
         )
         tex_file = Path(build_dir, "handbook.tex")
         with open(tex_file, "w+") as f:
             f.write(rendered_template)
-        shutil.copytree(f"{TEMPLATE_DIR}/content", f"{build_dir}/content")
+        if not Path(build_dir, "content").exists():
+            shutil.copytree(f"{TEMPLATE_DIR}/content", f"{build_dir}/content")
         subprocess.run(["pdflatex", f"-output-directory={build_dir}", str(tex_file)])
         subprocess.run(["makeindex", str(tex_file.with_suffix(".idx"))])
         subprocess.run(["pdflatex", f"-output-directory={build_dir}", str(tex_file)])
@@ -130,11 +132,14 @@ def process_papers(papers, root: Path, pax: bool):
     return id_to_paper, sorted(alphabetized_author_index.items())
 
 
-def process_program(program):
-    return sorted(program, key=lambda x: x["start_time"])
+def process_program_handbook(program):
+    sessions_by_date = defaultdict(list)
+    for session in program:
+        sessions_by_date[session["start_time"].date()].append(session)
+    return sorted(sessions_by_date.items())
 
 
-def process_program_by_page(program):
+def process_program_proceedings(program):
     """
     process_program organizes program sessions by date, and manually cuts
     program entries in order to avoid page overflow. This is done by assuming
@@ -144,12 +149,13 @@ def process_program_by_page(program):
     max_lines = 35
     paper_median_lines = 3
     header_lines = 2
-    dates = set()
-    for session in program:
-        dates.add(session["start_time"].date())
     sessions_by_date = defaultdict(list)
     for session in program:
-        sessions_by_date[session["start_time"].date()].append(session)
+        if "subsessions" in session:
+            for session in session["subsessions"]:
+                sessions_by_date[session["start_time"].date()].append(session)
+        else:
+            sessions_by_date[session["start_time"].date()].append(session)
     entries_by_date = {}
     for date, sessions in sessions_by_date.items():
         total_lines = 0
@@ -157,18 +163,22 @@ def process_program_by_page(program):
         current_page = []
         for session in sessions:
             table_entries = []
-            table_entries.append({
-                "type": "header",
-                "title": session["title"],
-                "start_time": session["start_time"],
-                "end_time": session["end_time"],
-            })
+            table_entries.append(
+                {
+                    "type": "header",
+                    "title": session["title"],
+                    "start_time": session["start_time"],
+                    "end_time": session["end_time"],
+                }
+            )
             if "papers" in session:
                 for paper_id in session["papers"]:
-                    table_entries.append({
-                        "type": "paper",
-                        "paper": paper_id,
-                    })
+                    table_entries.append(
+                        {
+                            "type": "paper",
+                            "paper": paper_id,
+                        }
+                    )
             # Split the table lines so that no page overflows.
             for entry in table_entries:
                 if entry["type"] == "header":
@@ -224,7 +234,9 @@ def load_config(config: str, root: Path, required=True):
     path = Path(root, f"{config}.yml")
     if not path.exists():
         if required:
-            raise ValueError(f"{config} is a required configuration but {config}.yml was not found")
+            raise ValueError(
+                f"{config} is a required configuration but {config}.yml was not found"
+            )
         return None
     with open(path, "r", encoding="utf-8") as f:
         return yaml.safe_load(f)
